Implemented AOE attacks
Created a function in the Enemy script that works differently with AOE attacks then the normal script. Updated the main tower script so that you can choose which type of attack the tower will do. AOE or normal. Main script for doing the AOE attack is in the bullet script.