Data for Rounds
Created the data needed to have multiple rounds. Updated the enemy and game manager so the enemies movement speed can be changed per round. Created a variable in the game manager script  can change the speed of when a new round begins. Created a few variables that enable the user to test the game without needing VR gear